+++ webpack.prod.js
@@ -7,7 +7,9 @@ module.exports = {
   entry: './src/index.js',
   output: {
     path: path.join(__dirname, 'dist'),
-    filename: 'index.js'
+    filename: 'index.js',
+    library: '@transferwise/promotion-service',
+    libraryTarget: 'umd',
   },
   module: {
     rules: [
@@ -21,8 +23,8 @@ module.exports = {
         test: /\.js$/,
         exclude: [path.resolve(__dirname, 'node_modules')],
         use: 'babel-loader',
-      }
-    ]
+      },
+    ],
   },
   plugins: [
     new webpack.optimize.ModuleConcatenationPlugin(),
@@ -34,32 +36,52 @@ module.exports = {
           comments: false,
           beautify: false,
         },
-      }
-    })
+      },
+    }),
   ],
   externals: [
     {
       angular: 'angular',
-      'angular-cookies': "'ngCookies'",
-      'angular-sanitize': "'ngSanitize'",
-      'angular-translate': "'pascalprecht.translate'",
-      'angular-translate-interpolation-messageformat': "''",
+      'angular-cookies': {
+        var: "'ngCookies'",
+        commonjs: 'angular-cookies',
+        commonjs2: 'angular-cookies',
+      },
+      'angular-sanitize': {
+        var: "'ngSanitize'",
+        commonjs: 'angular-sanitize',
+        commonjs2: 'angular-sanitize',
+      },
+      'angular-translate': {
+        var: "'pascalprecht.translate'",
+        commonjs: 'angular-translate',
+        commonjs2: 'angular-translate',
+      },
+      'angular-translate-interpolation-messageformat': {
+        var: "''",
+        commonjs: 'angular-translate-interpolation-messageformat',
+        commonjs2: 'angular-translate-interpolation-messageformat',
+      },
 
       underscore: '_',
       jquery: '$',
 
-      unorm: "''",
+      unorm: {
+        var: "''",
+        commonjs: 'unorm',
+        commonjs2: 'unorm',
+      },
 
       'tw-styleguide-components/dist/js/styleguide-components': {
         var: "''",
         commonjs: 'tw-styleguide-components/dist/js/styleguide-components',
-        commonjs2: 'tw-styleguide-components/dist/js/styleguide-components'
+        commonjs2: 'tw-styleguide-components/dist/js/styleguide-components',
       },
       '@transferwise/frontend-common/dist/tw-common': {
         var: "''",
-        commonjs: '@transferwise/frontend-common/dist/tw-common',
-        commonjs2: '@transferwise/frontend-common/dist/tw-common'
-      }
-    }
-  ]
+        commonjs: '@transferwise/frontend-common',
+        commonjs2: '@transferwise/frontend-common',
+      },
+    },
+  ],
 };