+++ CHANGELOG.md
@@ -1,3 +1,13 @@
+# v2.3.0
+## Adds UMD, change readme
+
+It is now possible to import this module from module-based systems by doing:
+`import Promotion from '@transferwise/promotion-service';`
+
+A named export, `promotionServiceName`, is also available, for injecting the service.
+
+In addition, a few errors in the readme have been fixed.
+
 # v1.0.3
 ## Change the angular module name to be consistent with the organisation naming convention